finished

+++ ft_funcs.h
@@ -21,7 +21,7 @@ int	ft_printc(char c);
 int	ft_prints(char *s);
 int	ft_printd(int n);
 int	ft_printu(unsigned int num);
-int	ft_printhex(unsigned int num, char xX);
+int	ft_printhex(unsigned long num, char xX);
 int	ft_printadd(unsigned long int num);
 
 #endif
